Refactor - Using cvcomputer instead of cvline in computer skills
Removed a lot of language with 3/4 TP practice (from 8 to 100 hours)
because it's irrelevant, using them would almost be like using a new language
and its clurreting the part to the point of unreadability. Also added stars
because when reading CV to hire an intern I just wanted to know the best
language and tool of an applicant and not waste time reading boring list.